+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+(no changes yet)
+
+## [0.3.1] - 2024-09-11
+
 ### tacacs-plus
 
 #### Removed
@@ -140,7 +144,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Shrubbery TACACS+ daemon]: https://shrubbery.net/tac_plus/
 
-[Unreleased]: https://github.com/cPacketNetworks/tacacs-plus-rs/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/cPacketNetworks/tacacs-plus-rs/compare/v0.3.1...HEAD
+[0.3.1]: https://github.com/cPacketNetworks/tacacs-plus-rs/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/cPacketNetworks/tacacs-plus-rs/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/cPacketNetworks/tacacs-plus-rs/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/cPacketNetworks/tacacs-plus-rs/compare/v0.2.0...v0.2.1